+++ bootstrap/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.4.0]
 
+- Add `/p2p/local/info` endpoint
 - Add webrtc support to bootstrap
 
 ## [0.3.0](https://github.com/availproject/avail-light/releases/tag/avail-light-bootstrap-v0.3.0) - 2024-09-27